Trying to interpolate into API field and getting
Ideally, upon the click of #supers i want to reassign the metricConven variable to a string; containing either 'imperial' or 'metric'. But there are issues in doing so that I don't know how to get around.

Trying to reassign it from within the toggleF_C function globally by reference via this.metricConven isn't working