(Update) API issues
API-issues:

At this point in the project, I am working with three 3rd Party APIs
- GeocoderAPI (Google Maps)
- PlacesAPI (Google Maps)
- Tomorrow API (Tomorrow.io)

As for the last two, there has been no issues, but a CORS same-origin policy has been preventing data from the PlacesAPI from being retrieved.

The options I'm considering are to either implement a proxy server to make my requests from or just completely disable the policy via Origin: "*".

I don't want this to slow me up though, so instead I'm going to suspend the use of PlacesAPI for the time being and resort to manually searching for cities without the aid of autocomplete.