Update(Upper-Main and Middle-Main)
Upper and middle main sections are not vertically aligned properly and the lower-main needs to inherit the properties of the first input field, then those same children should be block elements so they stack

After those comes the footer.
footer-col-1's children should be stacked and the pseudo-elements for the icons are to be added as well.

All footer column's fonts to be styled.